pyproject.toml
[tool.poetry]
name = "multicall"
version = "0.5.1"
description = "aggregate results from multiple ethereum contract calls with local caching"
authors = ["banteg <banteeg@gmail.com>", "parker burchett <parkerburchett@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
web3 = "^5.29.0"
eth_retry = "0.1.11"
aiohttp = "^3.8.1"
sqlitedict = "^2.1.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.3"
ruff = "^0.1.9"
pytest = "^6.2.5"
flake8 = "^4.0.1"
black = "^21.12b0"
click = "8.0.1"
joblib = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120

